Removed AWB internal buffer in favor of balancer
Instead of feeding items into an internal buffer then distributing them
to the craft grid, pipes can now access the craft grid directly and a
once per tick algorithm performs stack balancing on the grid to keep
things even.

This fixes weirdness with inserting items and the issue with overfilling
the buffer by feeding it multiple items at once (which resulted in items
spewing out of the AWB).